refs #7815
Perl scripts were written to modify LDAP in the case that a new version of an LDAP schema needs automated adjustments either before or after LDAP config generation

The Perl scripts were commented out.  We only used the perl scripts once, and if they need to be used in the future, their logic will need to be re-addressed.